event-logistics-swarm/requirements.txt
crewai
crewai-tools
langchain
openai
pandas
fastapi
uvicorn
python-dotenv